Removal of Symmetry flyout, split LayersPanel into more manageable parts.
